requirements.txt
absl-py==2.1.0
blinker==1.8.2
brax==0.10.5
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
contextlib2==21.6.0
dm-env==1.6
dm-tree==0.1.8
etils==1.7.0
Flask==3.0.3
Flask-Cors==4.0.1
flax==0.8.4
fsspec==2024.6.0
glfw==2.7.0
grpcio==1.64.1
gym==0.26.2
gym-notices==0.0.8
importlib_resources==6.4.0
itsdangerous==2.2.0
jax==0.4.29
jax-cuda12-pjrt==0.4.29
jax-cuda12-plugin==0.4.29
jaxlib==0.4.29
jaxopt==0.8.3
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
ml_collections==0.1.1
msgpack==1.0.8
mujoco==3.1.6
mujoco-mjx==3.1.6
nest-asyncio==1.6.0
numpy==2.0.0
nvidia-cublas-cu12==12.5.2.13
nvidia-cuda-cupti-cu12==12.5.39
nvidia-cuda-nvcc-cu12==12.5.40
nvidia-cuda-runtime-cu12==12.5.39
nvidia-cudnn-cu12==9.1.1.17
nvidia-cufft-cu12==11.2.3.18
nvidia-cusolver-cu12==11.6.2.40
nvidia-cusparse-cu12==12.4.1.24
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.5.40
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.16
packaging==24.1
pillow==10.3.0
protobuf==5.27.1
Pygments==2.18.0
PyOpenGL==3.1.7
pytinyrenderer==0.0.14
PyYAML==6.0.1
rich==13.7.1
scipy==1.13.1
six==1.16.0
tensorboardX==2.6.2.2
tensorstore==0.1.62
toolz==0.12.1
trimesh==4.4.1
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
